reafactor: Add useRtcClient hook
