+++ src/components/hooks/useRtcClient.ts
@@ -1,0 +1,22 @@
+import { useEffect, useReducer, useState } from 'react';
+
+import RtcClient from '../../utils/RtcClient';
+
+const useRtcClient = () => {
+  const [rtcClient, _setRtcClient] = useState<RtcClient | null>(null);
+  const [, forceRender] = useReducer((boolean) => !boolean, false);
+
+  const setRtcClient = (rtcClient: RtcClient) => {
+    _setRtcClient(rtcClient);
+    forceRender();
+  };
+
+  useEffect(() => {
+    const client = new RtcClient(setRtcClient);
+    client.setRtcClient();
+  }, []);
+
+  return rtcClient;
+};
+
+export default useRtcClient;
